requirements.txt
matplotlib
imagesize
numpy
scipy
Pillow
rasterio
torch
segmentation_models_pytorch
pytorch_lightning
torchmetrics==0.7.2
torchsummary
captum
Flask
Flask-Compress

